feat: make 404 page use ssr
