--- src/pages/404.tsx
@@ -1,38 +1,0 @@
-import React from 'react';
-import SEO from '../components/layout/seo';
-import { graphql, PageProps } from 'gatsby';
-import { SectionHeader } from '../components/content/section-header/section-header';
-import { useTranslation } from 'gatsby-plugin-react-i18next';
-import { Layout } from '../components/layout/layout';
-import { ContentBlockContainer } from '../components/layout/content-block-container';
-
-const NotFoundPage = ({ location }: PageProps): JSX.Element => {
-  const { t } = useTranslation();
-
-  return (
-    <>
-      <SEO title="404: Not found | Satellytes" location={location} />
-      <Layout light={true}>
-        <ContentBlockContainer>
-          <SectionHeader headline={'404'}>{t('404')}</SectionHeader>
-        </ContentBlockContainer>
-      </Layout>
-    </>
-  );
-};
-
-export default NotFoundPage;
-
-export const NotFoundPageQuery = graphql`
-  query ($language: String!) {
-    locales: allLocale(filter: { language: { eq: $language } }) {
-      edges {
-        node {
-          ns
-          data
-          language
-        }
-      }
-    }
-  }
-`;
